backend/requirements.txt
sqlmodel
pydantic
python-jose[cryptography]
fastapi[standard]
pydantic-settings
psycopg2-binary
redis
uvicorn